fix(web): keep write-lock hooks off the shortcut import graph
useAppShortcut pulled useAppAccess and billing-preview toasts at boot, so
AppConfigApi was still undefined when Google availability called .get.
